Created file
A script that automates the process of moving files from a source directory to a destination directory. The script takes the information about the source and destination directories and file names from a worksheet in an Excel file. The script starts by defining a worksheet variable and setting it to the worksheet that contains the data. Then it finds the last row of data in column A and uses that to determine how many files need to be moved.

The script then starts a loop that goes through each row of data starting from the second row. In each iteration of the loop, the script creates the full path of the source file and the destination file by concatenating the directory and file name information from columns A, B, C, and D.

Next, it checks if the destination folder exists, if it doesn't exist it creates the folder. After that, it checks if the source file exists, if it doesn't exist it writes an error message in column E saying 'Source file not found' and continues to the next iteration of the loop without moving the file. If the source file exists it uses the FileCopy method to copy the file from the source location to the destination location, and uses the Kill method to delete the original file from the source location. It also keeps track of the number of files moved.

Once the loop is finished, the script calculates the time it took to move the files by getting the current time in seconds before the loop starts and after the loop ends and calculates the total time by subtracting the start time from the end time and finally, it displays a message box that shows how many files were moved and how long the whole process took.

Please note that the script assumes that the data is in a worksheet named "Sheet1" and the data is starting from the second row, it is also important to note that the script will overwrite any existing files in the destination directory with the same name.